requirements.txt
llama-index
flask
torch
torchvision
torchaudio
pypdf
nbconvert
transformers
llama-cpp-python
python-dotenv
